adding null check
I'm not well versed in this space, but for some reason this line blows up when using the @iarna/rtf-to-html package. I'm parsing through sometimes large rtf files and I haven't figured out what in the files are causing this, but I can send a sample file if you'd like to see the error replicated.